Novas versões.

+++ matriz/src/criarMatriz.c
@@ -15,18 +15,18 @@ int main(int argc, char *argv[]) {
 
    // Faixa de valores aleatórios
    srand((unsigned int)time(NULL));
-   float faixaAleatoria = 100.00;
+   double faixaAleatoria = 100.0000;
 
    // Escreve a ordem.
    fprintf(arquivo, "%d\n", linhas);
    fprintf(arquivo, "%d\n", colunas);
    for (int i = 0; i < linhas; i++) {
       for (int j = 0; j < colunas; j++) {
-         float valorAleatorio = ((float)rand()/(float)(RAND_MAX)) * faixaAleatoria;
+         double valorAleatorio = ((double)rand()/(float)(RAND_MAX)) * faixaAleatoria;
          if (j < colunas - 1)
-            fprintf(arquivo, "%05.2f:", valorAleatorio);
+            fprintf(arquivo, "%05.4f:", valorAleatorio);
 	 else 
-            fprintf(arquivo, "%05.2f", valorAleatorio);
+            fprintf(arquivo, "%05.4f", valorAleatorio);
       }
       fprintf(arquivo, "\n");
    }
